# feat(dnsquery): implement HTTP client caching for DoH servers and prewarm connections