DoH cached values were HIGHER than uncached with extremely high standard deviation. This is illogical since cached DNS lookups should be faster as reported in the issue #23. 


The issue was caused by HTTP connection establishment overhead biasing the results:
- New HTTP client created per query → Limited connection pooling effectiveness
- Cached queries executed FIRST → They paid the TLS handshake cost
- Uncached queries executed LATER → They reused established connections
- Result: Connection setup time made cached queries appear slower


I verified this with a proof-of-concept test showing:
- Without client reuse: First query 42ms, erratic timing
- With client reuse: All queries consistent ~10ms


The solution has two parts:
- HTTP Client Pooling: Created persistent HTTP clients for DoH servers, stored in the Benchmarker struct and reused across all queries
- Connection Pre-warming: Added a pre-warming phase that establishes connections to DoH, DoT, and TCP servers BEFORE starting the benchmark, ensuring connection setup doesn't bias results

The result:
- pre-fix
```
Cloudflare DoH: Cached 34.7ms (±25.7ms) > Uncached 13.6ms (±0.2ms) 
Quad9 DoH:      Cached 21.0ms (±18.9ms) > Uncached 10.8ms (±0.5ms) 
```
- post-fix:
```
Cloudflare DoH: Cached 10.0ms (±0.6ms) < Uncached 13.3ms (±0.5ms) 
Quad9 DoH:      Cached  8.0ms (±0.3ms) < Uncached 14.3ms (±4.4ms) 
```